Improve add property
